After blood, sweat, and tears I'm done

+++ proj2/Tuple.cpp
@@ -16,16 +16,16 @@ string Tuple::toString(const Scheme& scheme) const {
     vector<string> schemeNonConstants;
     int numberOfGoodIterations = 0;
     // counting how many are constants
-    for (int nameValue = 0; nameValue < scheme.names.size(); nameValue++) {
-        if ((scheme.names.at(nameValue)).find("'") == -1) {
+    for (int nameValue = 0; unsigned(nameValue) < scheme.names.size(); nameValue++) {
+        if (int((scheme.names.at(nameValue)).find("'")) == -1) {
             schemeNonConstants.push_back(scheme.names.at(nameValue));
         }
     }
     int sizePrint = 0;
     vector<string> visitedAlreadyPartTwo;
     // checking how big the list is I need to print
-    for (int nameValue = 0; nameValue < scheme.names.size(); nameValue++) {
-        if ((scheme.names.at(nameValue)).find("'") == -1) {
+    for (int nameValue = 0; unsigned(nameValue) < scheme.names.size(); nameValue++) {
+        if (int((scheme.names.at(nameValue)).find("'")) == -1) {
             vector<string>::iterator it = find(visitedAlreadyPartTwo.begin(), visitedAlreadyPartTwo.end(), scheme.names.at(nameValue));
             if (it == visitedAlreadyPartTwo.end()) {
                 sizePrint++;
@@ -34,8 +34,8 @@ string Tuple::toString(const Scheme& scheme) const {
         visitedAlreadyPartTwo.push_back(scheme.names.at(nameValue));
     }
 
-    for (int nameValue = 0; nameValue < scheme.names.size(); nameValue++) {
-        if ((scheme.names.at(nameValue)).find("'") == -1) {
+    for (int nameValue = 0; unsigned(nameValue) < scheme.names.size(); nameValue++) {
+        if (int((scheme.names.at(nameValue)).find("'")) == -1) {
             if (numberOfGoodIterations == 0) {
                 out << "  ";
             }
